refactor: rename Network to NetworkUtils
Renamed the Network class to NetworkUtils for clarity. Updated all references in the login and signup services, as well as validation checks for network connectivity. This improves code readability and maintains consistency across the application.
